refactor direct-mount option handling to reuse functions from pure-rust impl and removed handling of CUSTOM options
